rolling time

+++ database/migrations/2026_03_26_155819_add_double_stamp_columns_to_results_and_totals.php
@@ -19,7 +19,7 @@ return new class extends Migration
 
         Schema::table('player_tournament_total', function (Blueprint $table) {
             $table->integer('double_stamp_landed')->default(0)->after('total');
-            $table->decimal('double_stamp_total', 10, 2)->default(0)->after('double_stamp_landed');
+            $table->integer('double_stamp_total')->default(0)->after('double_stamp_landed');
         });
     }
